Move slash-command tests into a dedicated file
Pull the 7 slash-command tests added in this branch out of the inline
`app::tests` and `commands::tests` modules into a single
`slash_command_tests.rs`. Declared via `#[path]` as a child of `app` (not
the crate root) so it keeps access to App's private dispatch methods, and
reuses `app::tests::test_app` (now `pub(super)`) instead of duplicating the
14-arg App::new wiring.

Pure refactor — same tests, same assertions, all green.
